requirement.txt
altgraph==0.17
future==0.18.2
pefile==2019.4.18
Pillow==8.2.0
pyinstaller-hooks-contrib==2021.1
PyQt5==5.15.4
PyQt5-Qt5==5.15.2
PyQt5-sip==12.8.1
PyQt5-stubs==5.15.2.0
pywin32-ctypes==0.2.0
